Move `SerializableMessage` under `type` package
