+++ base/src/main/java/io/spine/type/KnownMessage.java
@@ -27,7 +27,6 @@
 package io.spine.type;
 
 import com.google.errorprone.annotations.Immutable;
-import io.spine.base.SerializableMessage;
 
 /**
  * A message which was generated by Spine Model Compiler and is available
